Move page specific components into pages/ folder
