+++ packages/site-client/src/entrypoints/element.ts
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import '../components/wco-element-page.js';
-import type {ElementData} from '../components/wco-element-page.js';
+import '../pages/element/wco-element-page.js';
+import type {ElementData} from '../pages/element/wco-element-page.js';
 import {html} from 'lit';
 
 export const renderElementPage = (elementData: ElementData) =>
